Compile-time bugzilla has been felled. And now, the run-time darkness holds dominion, black as death.
Former-commit-id: 92945c1b61213712e538c817732c8a011c592c5d

+++ source/ArgumentParser.h
@@ -12,12 +12,12 @@
 #include <string>
 #include <vector>
 #include <map>
-
+#include "Settings.h"
 using namespace std;
 
 namespace ArgumentParser
 {
-	enum Types { NONE, INT, DOUBLE, CHAR, STRING };
+	enum Types { NONE, INT, REAL, CHAR, STRING };
 	
 	struct OptionStruct
 	{
